+++ shared/src/commonMain/kotlin/ui/composables/DCQLCredentialQuerySubmissionSelection.kt
@@ -8,13 +8,13 @@ import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.dp
 import at.asitplus.openid.dcql.DCQLCredentialSubmissionOption
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
-import at.asitplus.wallet.lib.agent.validation.CredentialFreshnessSummary
+import ui.models.CredentialFreshnessSummaryUiModel
 
 
 @Composable
 fun DCQLCredentialQuerySubmissionSelection(
     selectionOptions: List<DCQLCredentialSubmissionOption<SubjectCredentialStore.StoreEntry>>,
-    checkFreshness: suspend (SubjectCredentialStore.StoreEntry) -> CredentialFreshnessSummary?,
+    checkCredentialFreshness: suspend (SubjectCredentialStore.StoreEntry) -> CredentialFreshnessSummaryUiModel,
     currentlySelectedOptionIndex: Int?,
     onChangeSelection: (Int?) -> Unit,
     decodeToBitmap: (ByteArray) -> ImageBitmap?,
@@ -27,7 +27,7 @@ fun DCQLCredentialQuerySubmissionSelection(
         selectionOptions.forEachIndexed { index, option ->
             val isSelected = index == currentlySelectedOptionIndex
             DCQLCredentialQuerySubmissionSelectionOption(
-                checkFreshness = checkFreshness,
+                checkCredentialFreshness = checkCredentialFreshness,
                 isSelected = isSelected,
                 onToggleSelection = { onChangeSelection(if (isSelected) null else index) },
                 decodeToBitmap = decodeToBitmap,