+++ shared/src/commonMain/kotlin/ui/models/TimelinessIndicatorUiModel.kt
@@ -1,0 +1,8 @@
+package ui.models
+
+interface TimelinessIndicatorUiModel {
+    val isExpired: Boolean
+    val isNotYetValid: Boolean
+    val isTimely: Boolean
+        get() = !isNotYetValid && !isExpired
+}